==> CR/0_data_prep.ipynb <==
!pip3 install gdown > /dev/null

import gdown
url = 'https://drive.google.com/u/0/uc?export=download&confirm=1d1e&id=1IXzOffSk-jJHMYhBJdfrHRegKw1ycdr9'
output = 'data.csv'
gdown.download(url, output, quiet=False) 

import pandas as pd

data = pd.read_csv("data.csv")

data.info()

data.head()
